+++ blockchain-JS/.eslintrc.yaml
@@ -1,9 +1,13 @@
 {   
     "extends": "eslint:recommended",
-    "env": { "es6": true },
+    "env": { 
+        "es6": true,
+        "browser": true,
+        "node": true,
+        },
     "rules": {
         "semi": ["error", "always"],
-        "quotes": ["error", "double"]
+        "quotes": ["error", "single"]
     },
     "overrides": [
         {
