+++ blockchain-JS/app.js
@@ -2,22 +2,21 @@ const express = require('express');
 const expressLayouts = require('express-ejs-layouts');
 const path = require('path');
 
-// Serup Server
+// Setup Server
 const app = express();
 
 
-//do init
-const {Miner} = require('./controller/miner')
+// Initiating
+const {Miner} = require('./source/components/miner')
 
 const miner = new Miner();
 miner.initTransactions();
 
-// Setup views
+// Setup Views
 app.set('view engine', 'ejs');
-app.set('views', __dirname + '/views');
+app.set('views', __dirname + '/source/views');
 app.use(expressLayouts);
-require('./routes/app.routes.js')(app, miner);
-app.use(express["static"](path.join(__dirname, '/css')));
+require('./source/routes/app.routes.js')(app, miner);
+app.use(express["static"](path.join(__dirname, '/source/css')));
+
 module.exports = app;
-// Start server
-//app.listen(port, () => console.log(`Magic happens on port http://localhost:${port}!`))
